moved some code from main to two new modules
