+++ BigIntCalculator/evalExpr.cpp
@@ -19,7 +19,6 @@ along with Alpertron Calculators.If not, see < http://www.gnu.org/licenses/>.
 // declarations for external function
 
 void biperm(int n, Znum &result);  
-Znum llt(const Znum &p);
 size_t inverseTotient(__int64 n, std::vector<unsigned __int64> **result, bool debug,
     int level, bool dump);
 
@@ -51,6 +50,60 @@ struct
 
 /* list of operators, arranged in order of priority, order is not exactly the
 same as C or Python. Followed by list of function codes*/
+/* operators and functions include:
+symbol		meaning					operator Priority
+n!			factorial               0
+n!..!	    multi-factorial         0
+p#			primorial               0
+            (product of all primes
+            less or equal than p)
+
+-			unary minus				1    (right to left)
+NOT			bitwise operators		1    (right to left)
+** or ^		Exponentiate			2    (right to left)
+*			multiply				3
+/			divide					3
+%			modulus (remainder)		3
+C			binomial coefficient	4
++			add						5
+-			subtract				5
+n SHL b		shift n left b bits		6
+n SHR b		shift n right b bits	6
+
+comparison operators
+>			return zero for false 	7
+>=			and -1 for true			7
+<									7
+<=									7
+==									8
+!=			not equal				8
+
+AND			bitwise operators		9
+OR									11
+XOR									10
+
+
+________________ Functions include ___________________________________
+GCD(a, b, ....)
+MODPOW(m,n,r): finds m^n modulo r
+MODINV(m,n): inverse of m modulo n, only valid when gcd(m,n)=1
+SUMDIGITS(n,r): Sum of digits of n in base r.
+NUMDIGITS(n,r): Number of digits of n in base r
+REVDIGITS(n,r): finds the value obtained by writing backwards the digits of n in base r.
+ISPRIME(n)
+F(n):		Fibonacci
+L(n):		Lucas
+P(n):		Unrestricted Partition Number (number of decompositions of n
+            into sums of integers without regard to order)
+PI(n):		Number of primes <= n
+N(n):		Next probable prime after n
+B(n):		Previous probable prime before n
+Totient(n): finds the number of positive integers less than n which are relatively prime to n.
+NumDivs(n): Number of positive divisors of n either prime or composite.
+SumDivs(n): Sum of positive divisors of n either prime or composite.
+FactConcat(m,n): Concatenates the prime factors of n according to the mode expressed in m
+
+*/
 enum class opCode {
     fact        = 21,	// !   factorial 
     prim        = 23,	// #   primorial
@@ -250,7 +303,7 @@ const static struct oper_list operators[]{
 enum class types { Operator, func, number, comma, error, uservar, end };
 
 struct token {
-    types typecode;
+    types typecode; /* type of token (operator, function, number, etc) */
     int function;   /* contains function code index, when typecode = func 
                        contains operator index when typecode = Operator */
     opCode oper;    /* contains operator value, only when typecode = Operator or func */
@@ -866,6 +919,183 @@ static Znum primRoot(const Znum &num) {
     }
 
     return -1;  /* should never reach this point */
+}
+
+// find leftmost 1 bit of number. Bits are numbered from 63 to 0
+// An intrinsic is used that gets the bit number directly.
+// If n = zero the value returned is 0, although bit 0 is zero
+static int leftBit(unsigned __int64 n) {
+    int bit = 0, nz = 1;
+
+    nz = _BitScanReverse64((DWORD*)&bit, n);  // this intrinsic should be expanded inline
+    // to use the fastest available machine instructions
+    if (nz) return (int)bit;
+    else return 0;   // n actually contains no 1-bits.
+
+}
+
+/* n = 2^p -1. Increment lltTdivCnt & return 0 if trial division finds a factor,
+   otherwise return 1 */
+static int lltTrialDiv(const Znum& n, const long long p) {
+    Znum tmp, ncopy, limit;
+    const int maxlimit = 1000000;  /* Limits the number of iterations for trial
+                                     division. Experiments indicate that increasing
+                                     this limit slows llt down overall. */
+    bool composite = false;
+
+    /* do trial division */
+    ncopy = n;
+
+    /*  n is a mersenne number calculated as (2^p)-1 and p is a prime number.
+        The factors of n must be in the form of q = 2ip+1, where q < n.
+        2ip+1 < n therefore
+        i < (n-1)/2p */
+        //limit = (sqrt(n) - 1) / (2 * p) + 1;   /* mpz_sqrt is probably faster */
+    Znum nm1 = n - 1;
+    mpz_sqrt(ZT(limit), ZT(nm1));
+    /* if we find all factors < sqrt(n) the residue is the one remaining factor */
+
+    if (limit > maxlimit) {
+        /* hit this limit if p >= 41 */
+        if (verbose > 1)
+            gmp_printf("limit reduced from %Zd to %lld\n", limit, maxlimit);
+        limit = maxlimit;   // larger limit would take too long (also encounter > 64-bit integers)
+    }
+
+    long long ll_limit = MulPrToLong(limit);
+    Znum rem;
+    for (long long i = 1; i <= ll_limit; i++) {
+        long long d = 2 * i * p + 1;
+        //if (n%d == 0)
+        long long r = mpz_tdiv_r_ui(ZT(rem), ZT(n), d);
+        if (r == 0) {
+            composite = true;
+            ncopy /= d;
+            if (verbose > 0)
+                printf_s("2*%lld*p+1 (= %lld) is a factor\n", i, d);
+            else {
+                lltTdivCnt++;  /* increment counter */
+                return 0;  /* not prime; return as soon as we know this */
+            }
+        }
+    }
+
+    if (composite) {
+        if (ncopy > 1 && verbose > 0)
+            gmp_printf("%Zd is a factor\n", ncopy);
+        lltTdivCnt++;   /* increment counter */
+        return 0; /* not prime */
+    }
+    else {
+        if (verbose > 0)
+            printf_s("%s No factors found by trial division to %d bits, start Lucas-Lehmer test \n",
+                myTime(), leftBit(2 * ll_limit * p + 1) + 1);
+        return 1;  /* no factor found */
+    }
+}
+
+/* get remainder when num is divided by m = 2^p-1. The obscure method used
+avoids division by 2^p-1, and divides by 2^p instead which is much faster,
+as it is equivalent to a right shift by p bits.
+For efficiency both p and m are given as parameter values although a cleaner
+interface would just supply p and let m be recalculated. */
+static void getrem(Znum& rem, const Znum& num, const long long p, const Znum& m) {
+    static Znum mod2p, q;  /* static for efficiency */
+    mod2p = num & m;     // mod2p = num%(2^p) = num -q*2^p
+    mpz_tdiv_q_2exp(ZT(q), ZT(num), p);  // q = num/(2^p)
+    rem = q + mod2p;                    //rem = num-q*(2^p-1)
+    while (rem >= m)
+        rem -= m;               // reduce rem if necessary until 0 <= rem < (2^p-1)
+#ifdef _DEBUG
+    /* demonstrate that the result is correct */
+    //Znum  temp;
+    //mpz_tdiv_r(ZT(temp), ZT(num), ZT(m));
+    //assert(temp == rem);
+#endif
+}
+
+/* perform Lucas-Lehmer test. Return 0 if 2^p-1 is composite, 1 if prime
+see https://en.wikipedia.org/wiki/Lucas%E2%80%93Lehmer_primality_test
+also https://oeis.org/A000043
+and https://www.mersenne.org/primes/
+the largest mersenne prime found so far is p=82589933
+p must not be too large, otherwise it would take centuries to get a result.
+This code is based on the llt function in YAFU, but has been modified to make
+it much faster.
+if verbose is > 0 some extra messages are output.
+If verbose is > 1 trial division is used to try to find factors */
+Znum llt(const Znum& p) {
+    int rv;
+    Znum n, tmp;
+    long long exp = MulPrToLong(p); /* exp = p */
+    clock_t t1, t2, t3;
+
+    /* 1st check if p is prime. */
+    rv = mpz_bpsw_prp(ZT(p));  /* returns 0, 1 or 2*/
+    /* rv is 1 if p is probably prime, or 0 if p is definitely composite.*/
+    if (rv == 0) {
+        lltTdivCnt++;  // count this result as found by trial division
+        return 0;      // if p is composite, 2^p -1 is composite.
+    }
+    if (p <= 7) {
+        lltPriCnt++;  // count as found by Lucas-Lehmer test
+        return 1;    /* 1st mersenne number that is not prime is p=11 */
+    }
+
+    n = 1;
+    mpz_mul_2exp(ZT(n), ZT(n), exp);   // n = 2^p
+    n -= 1;                            // n = 2^p-1
+
+    /* experiment - is mpz-probab_prime_p faster overall than trial division?
+     Results indicate yes, definitely. Maybe because it is >99% effective in screening
+     out composites with just 2 tests so that llt is only neeeded as confirmation.
+     Update 11/7/2021; using neither trial division nor  mpz-probab_prime_p is fastest.
+     LL test is in fact faster than miller-rabin even when MR is limitied to 2 rounds. */
+    if (verbose > 1) {
+        t1 = clock();
+        /* use trial division if verbose > 1 because this can give some factors,
+        whereas the other tests only show whether n is prime or composite */
+        int rv = lltTrialDiv(n, exp);
+        t2 = clock();
+        printf_s("time used by trial division = %.2f sec \n", (double)(t2 - t1) / CLOCKS_PER_SEC);
+        if (rv == 0) {
+            std::cout << myTime() << " 2^" << exp << " -1 is not prime *** \n";
+            return 0;   /* factor found -> n is composite */
+        }
+    }
+
+    /*do the ll test */
+    tmp = 4;
+    int nchars = 0;
+    if (verbose > 0)
+        t2 = clock();
+    for (long long i = 0; i < exp - 2; i++) {
+        tmp *= tmp;
+        tmp -= 2;
+        //tmp %= n;
+        getrem(tmp, tmp, exp, n);  /* get remainder of division by n */
+        if (verbose > 0 && (i > 0) && (i & 0x7ff) == 0) { /* 7ff = 2047, print msg every 2048 iterations */
+            for (int j = 0; j < nchars; j++)
+                printf_s("\b");    // erase previous output
+            nchars = printf_s("%s llt iteration %lld %.2f%% complete", myTime(), i,
+                (double)i * 100.0 / (exp - 2));
+            fflush(stdout);
+        }
+    }
+    if (verbose > 0) {
+        t3 = clock();
+        printf_s(lang ? "\ntiempo usado por llt = %.2f sec \n" : "\ntime used by llt = %.2f sec \n",
+            (double)(t3 - t2) / CLOCKS_PER_SEC);
+    }
+
+    if (tmp == 0) {
+        lltPriCnt++;  /* increment counter */
+        return 1;             // prime
+    }
+    else {
+        lltCmpCnt++;   /* increment counter */
+        return 0;            // composite
+    }
 }
 
 
@@ -1604,58 +1834,6 @@ static void printTokens(const std::vector <token> expr) {
 }
 
 /* evaluate expression and return value in Result
-operators and functions include:
-symbol		meaning					operator Priority
-n!			factorial               0
-n!..!	    multi-factorial         0
-p#			primorial               0
-            (product of all primes
-            less or equal than p)
-
--			unary minus				1    (right to left)
-NOT			bitwise operators		1    (right to left)
-** or ^		Exponentiate			2    (right to left)
-*			multiply				3
-/			divide					3
-%			modulus (remainder)		3
-C			binomial coefficient	4
-+			add						5
--			subtract				5
-n SHL b		shift n left b bits		6
-n SHR b		shift n right b bits	6
-
-comparison operators
->			return zero for false 	7
->=			and -1 for true			7
-<									7
-<=									7
-==									8
-!=			not equal				8
-
-AND			bitwise operators		9
-OR									11
-XOR									10
-
-
-________________Functions___________________________________
-GCD(a, b)
-MODPOW(m,n,r): finds m^n modulo r
-MODINV(m,n): inverse of m modulo n, only valid when gcd(m,n)=1
-SUMDIGITS(n,r): Sum of digits of n in base r.
-NUMDIGITS(n,r): Number of digits of n in base r
-REVDIGITS(n,r): finds the value obtained by writing backwards the digits of n in base r.
-ISPRIME(n)
-F(n):		Fibonacci
-L(n):		Lucas
-P(n):		Unrestricted Partition Number (number of decompositions of n
-            into sums of integers without regard to order)
-PI(n):		Number of primes <= n
-N(n):		Next probable prime after n
-B(n):		Previous probable prime before n
-Totient(n): finds the number of positive integers less than n which are relatively prime to n.
-NumDivs(n): Number of positive divisors of n either prime or composite.
-SumDivs(n): Sum of positive divisors of n either prime or composite.
-FactConcat(m,n): Concatenates the prime factors of n according to the mode expressed in m
 
 Hexadecimal numbers are preceded by 0X or 0x
 */
@@ -1871,7 +2049,7 @@ If there is more than one number on the stack at the end, or at any time there
 are not enough numbers on the stack to perform an operation an error is reported.
 (this would indicate a syntax error not detected earlier) 
 If the final operation is a function call that returns multiple values,
-multiValue is set to true, otherwise it is false*/
+multiValue is set to true, otherwise it is set to false*/
 static retCode evalExpr(const std::vector<token> &rPolish, Znum & result, bool *multiV) {
     std::stack <token> nums;   /* this stack holds both numbers and user variables */
     Znum val;
@@ -2033,6 +2211,56 @@ retCode ComputeExpr(const std::string &expr, Znum &Result, int &asgCt, bool *mul
             //putchar('\n');
         }
     }
+
+    return rv;
+}
+
+/* evaluate 1 or more expressions, separated by commas */
+retCode ComputeMultiExpr(std::string expr, Znum result) {
+    std::string subExpr;
+    retCode rv = retCode::EXPR_OK;
+    size_t subStart = 0, subEnd;
+    int bc = 0;   /* bracket count */
+    int asgCt = 0;   /* number of assignment operators */
+    while (subStart < expr.size()) {
+        bc = 0;
+        /* find  separator or end of text*/
+        for (subEnd = subStart + 1; subEnd < expr.size(); subEnd++) {
+            if (expr[subEnd] == '(')
+                bc++;
+            if (expr[subEnd] == ')')
+                bc--;
+            if (bc == 0 && expr[subEnd] == ',')
+                break;
+        }
+        if (bc != 0)
+            return retCode::PAREN_MISMATCH;
+        subExpr = expr.substr(subStart, subEnd - subStart);
+        removeInitTrail(subExpr);  /* remove initial & trailing blanks */
+        removeIntSpace(subExpr);   /* remove spaces between digits */
+        rv = ComputeExpr(subExpr, result, asgCt);
+        if (rv != retCode::EXPR_OK) {
+            textError(rv);   // invalid expression; print error message
+            return rv;
+        }
+        else {
+            if (asgCt == 0)
+                std::cout << " = ";
+            else {
+                /* print names of variables assigned values */
+                for (size_t ix = 0; ix < subExpr.size() && asgCt > 0; ix++) {
+                    putchar(subExpr[ix]);
+                    if (subExpr[ix] == '=')
+                        asgCt--;
+                }
+            }
+
+            ShowLargeNumber(result, groupSize, true, hexPrFlag);   // print value of expression
+            std::cout << '\n';
+
+        }
+        subStart = subEnd + 1;  /* move past , */
+    } /* end of while loop */
 
     return rv;
 }
