+++ BigIntCalculator/main.cpp
@@ -15,9 +15,6 @@ along with Alpertron Calculators.  If not, see <http://www.gnu.org/licenses/>.
 #pragma fenv_access (on)  /* allow access to floating point environment */
 
 #include "pch.h"
-#include <fcntl.h>
-#include <io.h>
-#include <windows.h>
 #include <strsafe.h>
 
 #include <Mmsystem.h >   // for sound effects
@@ -35,8 +32,6 @@ processorArchitecture='*' publicKeyToken='6595b64144ccf1df' language='*'\"")
 #ifdef BIGNBR
 extern Znum zR, zR2, zNI, zN;
 #endif
-
-extern long long SFhitcount, SFmisscount;
 
 EXTERN_C IMAGE_DOS_HEADER __ImageBase;
 #define HINST_THISCOMPONENT ((HINSTANCE)&__ImageBase)
@@ -65,6 +60,9 @@ bool *primeFlags = NULL;
 unsigned long long *primeList = NULL;
 unsigned int prime_list_count = 0;
 unsigned long long int primeListMax = 0;
+// machine info
+double MEAS_CPU_FREQUENCY;
+char CPU_ID_STR[80] = { '\0' };
 
 std::vector <std::string> exprList;  /* expressions stored here as text once
                                        they are validated */
@@ -127,18 +125,6 @@ const char * myTime(void) {
     return timestamp;
 }
 
-// find leftmost 1 bit of number. Bits are numbered from 63 to 0
-// An intrinsic is used that gets the bit number directly.
-// If n = zero the value returned is 0, although bit 0 is zero
-static int leftBit(unsigned __int64 n) {
-    int bit = 0, nz = 1;
-
-    nz = _BitScanReverse64((DWORD*)&bit, n);  // this intrinsic should be expanded inline
-            // to use the fastest available machine instructions
-    if (nz) return (int)bit;
-    else return 0;   // n actually contains no 1-bits.
-    
-}
 
 /* Convert number to hexdecimal. Ensure that if number is negative the leftmost
 bit of the most significant digit is set, and conversely, if the number is positive
@@ -273,169 +259,7 @@ long long lltCmpCnt = 0;   /* count no of Mersenne numbers determined to be comp
 long long lltPriCnt = 0;   /* count no of Mersenne numbers determined to be prime
                            using Lucas-Lehmer test */
 
-/* n = 2^p -1. Increment lltTdivCnt & return 0 if trial division finds a factor, 
-   otherwise return 1 */
-static int lltTrialDiv(const Znum &n, const long long p) {
-    Znum tmp, ncopy, limit;
-    const int maxlimit = 1000000;  /* Limits the number of iterations for trial
-                                     division. Experiments indicate that increasing
-                                     this limit slows llt down overall. */
-    bool composite = false;
 
-    /* do trial division */
-    ncopy = n;
-
-    /*  n is a mersenne number calculated as (2^p)-1 and p is a prime number.
-        The factors of n must be in the form of q = 2ip+1, where q < n.
-        2ip+1 < n therefore
-        i < (n-1)/2p */
-        //limit = (sqrt(n) - 1) / (2 * p) + 1;   /* mpz_sqrt is probably faster */
-    Znum nm1 = n - 1;
-    mpz_sqrt(ZT(limit), ZT(nm1));
-    /* if we find all factors < sqrt(n) the residue is the one remaining factor */
-
-    if (limit > maxlimit) {
-        /* hit this limit if p >= 41 */
-        if (verbose > 1)
-            gmp_printf("limit reduced from %Zd to %lld\n", limit, maxlimit);
-        limit = maxlimit;   // larger limit would take too long (also encounter > 64-bit integers)
-    }
-
-    long long ll_limit = MulPrToLong(limit);
-    Znum rem;
-    for (long long i = 1; i <= ll_limit; i++) {
-        long long d = 2 * i*p + 1;
-        //if (n%d == 0)
-        long long r = mpz_tdiv_r_ui(ZT(rem), ZT(n), d);
-        if (r == 0) {
-            composite = true;
-            ncopy /= d;
-            if (verbose > 0)
-                printf_s("2*%lld*p+1 (= %lld) is a factor\n", i, d);
-            else {
-                lltTdivCnt++;  /* increment counter */
-                return 0;  /* not prime; return as soon as we know this */
-            }
-        }
-    }
-
-    if (composite) {
-        if (ncopy > 1 && verbose > 0)
-            gmp_printf("%Zd is a factor\n", ncopy);
-        lltTdivCnt++;   /* increment counter */
-        return 0; /* not prime */
-    }
-    else {
-        if (verbose > 0)
-        printf_s("%s No factors found by trial division to %d bits, start Lucas-Lehmer test \n",
-            myTime(), leftBit(2 * ll_limit * p + 1) + 1);
-        return 1;  /* no factor found */
-    }
-}
-
-/* get remainder when num is divided by m = 2^p-1. The obscure method used 
-avoids division by 2^p-1, and divides by 2^p instead which is much faster,
-as it is equivalent to a right shift by p bits. 
-For efficiency both p and m are given as parameter values although a cleaner
-interface would just supply p and let m be recalculated. */
-static void getrem(Znum &rem, const Znum &num, const long long p, const Znum &m) {
-    static Znum mod2p, q;  /* static for efficiency */
-    mod2p = num & m;     // mod2p = num%(2^p) = num -q*2^p
-    mpz_tdiv_q_2exp(ZT(q), ZT(num), p);  // q = num/(2^p)
-    rem = q + mod2p;                    //rem = num-q*(2^p-1)
-    while (rem >= m)
-        rem -= m;               // reduce rem if necessary until 0 <= rem < (2^p-1)
-#ifdef _DEBUG
-    /* demonstrate that the result is correct */
-    //Znum  temp;
-    //mpz_tdiv_r(ZT(temp), ZT(num), ZT(m));
-    //assert(temp == rem);
-#endif
-}
-
-/* perform Lucas-Lehmer test. Return 0 if 2^p-1 is composite, 1 if prime  
-see https://en.wikipedia.org/wiki/Lucas%E2%80%93Lehmer_primality_test 
-also https://oeis.org/A000043 
-and https://www.mersenne.org/primes/ 
-the largest mersenne prime found so far is p=82589933 
-p must not be too large, otherwise it would take centuries to get a result.
-This code is based on the llt function in YAFU, but has been modified to make 
-it much faster. 
-if verbose is > 0 some extra messages are output.
-If verbose is > 1 trial division is used to try to find factors */
-Znum llt(const Znum &p) {
-    int rv;
-    Znum n, tmp; 
-    long long exp = MulPrToLong(p); /* exp = p */
-    clock_t t1, t2, t3;
-
-    /* 1st check if p is prime. */
-    rv = mpz_bpsw_prp(ZT(p));  /* returns 0, 1 or 2*/
-    /* rv is 1 if p is probably prime, or 0 if p is definitely composite.*/
-    if (rv == 0) {
-        lltTdivCnt++;  // count this result as found by trial division
-        return 0;      // if p is composite, 2^p -1 is composite.
-    }
-    if (p <= 7) {
-        lltPriCnt++;  // count as found by Lucas-Lehmer test
-        return 1;    /* 1st mersenne number that is not prime is p=11 */
-    }
-
-    n = 1;
-    mpz_mul_2exp(ZT(n), ZT(n), exp);   // n = 2^p
-    n -= 1;                            // n = 2^p-1
-
-    /* experiment - is mpz-probab_prime_p faster overall than trial division? 
-     Results indicate yes, definitely. Maybe because it is >99% effective in screening
-     out composites with just 2 tests so that llt is only neeeded as confirmation. 
-     Update 11/7/2021; using neither trial division nor  mpz-probab_prime_p is fastest.
-     LL test is in fact faster than miller-rabin even when MR is limitied to 2 rounds. */
-    if (verbose > 1) {
-        t1 = clock();
-        /* use trial division if verbose > 1 because this can give some factors, 
-        whereas the other tests only show whether n is prime or composite */
-        int rv = lltTrialDiv(n, exp);
-        t2 = clock();
-        printf_s("time used by trial division = %.2f sec \n", (double)(t2 - t1) / CLOCKS_PER_SEC);
-        if (rv == 0) {
-            std::cout << myTime() << " 2^" << exp << " -1 is not prime *** \n";
-            return 0;   /* factor found -> n is composite */
-        }
-    }
-    
-    /*do the ll test */
-    tmp = 4;
-    int nchars = 0;
-    if (verbose > 0) 
-        t2 = clock();
-    for (long long i = 0; i < exp - 2; i++) {
-        tmp *= tmp;
-        tmp -= 2;
-        //tmp %= n;
-        getrem(tmp, tmp, exp, n);  /* get remainder of division by n */
-        if (verbose > 0 && (i > 0) && (i & 0x7ff) == 0) { /* 7ff = 2047, print msg every 2048 iterations */
-            for (int j = 0; j < nchars; j++)
-                printf_s("\b");    // erase previous output
-            nchars = printf_s("%s llt iteration %lld %.2f%% complete", myTime(), i,
-                (double)i*100.0/(exp-2) );
-            fflush(stdout);
-        }
-    }
-    if (verbose > 0) { 
-        t3 = clock();
-        printf_s(lang? "\ntiempo usado por llt = %.2f sec \n": "\ntime used by llt = %.2f sec \n", 
-            (double)(t3 - t2) / CLOCKS_PER_SEC);
-    }
-
-    if (tmp == 0) {
-        lltPriCnt++;  /* increment counter */
-        return 1;             // prime
-    }
-    else {
-        lltCmpCnt++;   /* increment counter */
-        return 0;            // composite
-    }
-}
 
 
 /* get floor(sqrt(n))*/
@@ -679,7 +503,7 @@ static void removeBlanks(std::string &msg) {
 }
 
 /* remove initial & trailing spaces, tabs, etc from msg (\t, \r, \n, \v   and \f) */
-static void removeInitTrail(std::string &msg) {
+void removeInitTrail(std::string &msg) {
     while (!msg.empty() && isspace(msg.front())) {
         msg.erase(0, 1);      // remove 1st space character
     }
@@ -689,7 +513,7 @@ static void removeInitTrail(std::string &msg) {
 }
 
 /* remove any spaces between 2 digits, also multiple consecutive spaces reduced to 1 space */
-static void removeIntSpace(std::string& msg) {
+void removeIntSpace(std::string& msg) {
     if (msg.empty())
         return;
     for (int i = 1; i < (msg.size() - 1);) {
@@ -2144,1063 +1968,6 @@ static void processIni(const char * arg) {
     }
 }
 
-
-struct HelpDialogResp {
-    int radiobutton;
-    int Ok_Cancel;
-};
-
-HelpDialogResp hResp;  /* user choices in the help dialog box are saved here */
-
-/* process messages from dialog box for Help topics 
-returns TRUE(1) or FALSE(0) */
-INT_PTR helpDialogAct(HWND DiBoxHandle,
-    UINT message,
-    WPARAM additionalInfo1,
-    LPARAM additionalInfo2) {
-
-    int wpHi = HIWORD(additionalInfo1);     /* 0 or control notification code */
-    int wpLo = LOWORD(additionalInfo1);     /* menu or control identifier */
-
-    int button = 0;  /* set according to radio button selected */
-    //char butttext[][20] = { "HELP", "FUNCTION", "EXPRESSION", "OTHER",
-    //        "YAFU", "TEST", "MSIEVE", "BACKGROUND", "LOOP", "QMES",
-    //        "PARI", };
-
-    switch (message) { 
-
-    case WM_DESTROY:        /* 0x02 */
-    case  WM_MOVE:          /* 0x03 */
-    case WM_ACTIVATE:       /* 0x06 */
-    case WM_SETFOCUS:       /* 0x07 */
-    case WM_KILLFOCUS:      /* 0x08 */
-    case WM_PAINT:          /* 0x0f */
-    case WM_ERASEBKGND:     /* 0x14 */
-    case WM_SHOWWINDOW:     /* 0x18 */
-    case WM_ACTIVATEAPP:    /* ox1c */
-    case WM_SETCURSOR:      /* 0x20 */
-    case WM_MOUSEACTIVATE:  /* 0x21 */
-    case WM_GETMINMAXINFO:  /* 0x24 */
-    case WM_SETFONT:        /* 0x30 */
-    case WM_WINDOWPOSCHANGING:  /* 0x46 */
-    case WM_WINDOWPOSCHANGED:  /* 0x47 */
-    case WM_NOTIFY:            /* 0x4e */
-    case WM_HELP:              /* ox53 */
-    case WM_NOTIFYFORMAT:      /* 0x55 */
-    case WM_GETICON:           /* 0x7f */
-    case WM_NCDESTROY:         /* 0x82 */
-    case WM_NCHITTEST:     /* 0x84 */
-    case WM_NCPAINT:       /* 0x85 */
-    case WM_NCACTIVATE:    /* 0x86 */
-    case 0x90:             /* can't find any documentation for this */
-    case WM_NCMOUSEMOVE:   /* 0xa0 */
-    case WM_NCLBUTTONDOWN: /* 0xa1 */
-    case 0xae:             /* can't find any documentation for this */
-        return FALSE;
-
-    case WM_INITDIALOG:    /* 0x110 */ {
-        /* select 1st radio button */
-        BOOL rv = CheckRadioButton(DiBoxHandle, general, pari, general);
-        if (rv == FALSE) {
-            ErrorDisp(__FUNCTION__);
-        }
-        /* return true so system sets focus to 1st control */
-        return TRUE;
-    }
-
-    case WM_COMMAND: /* 0x111 control selected by user */
-        switch (wpLo) {  /* switch according to control selected */
-        case general:
-        case function:
-        case expression:
-        case Other:
-        case YAFU:
-        case Test:
-        case Msieve:
-        case Background:
-        case Loop:
-        case QMES:
-        case pari:
-            button = wpLo - general;  /* general -> 0, function -> 1 etc. */
-            hResp.radiobutton = button;
-            //if (verbose >0)
-            //    std::cout << "button = " << butttext[button] << '\n';
-            return FALSE;
-
-        case IDOK:       /* OK         */
-        case IDCANCEL:   /* cancel */
-            hResp.Ok_Cancel = wpLo;
-            EndDialog(DiBoxHandle, wpLo);  /* close dialog box */
-            return TRUE;
-
-        default:  /* unknown control*/
-            std::cout << "HelpDialog WM_COMMAND  wpHi = " << wpHi << " wpLo = " << wpLo
-                << " info2 = " << additionalInfo2 << '\n';
-        }
-        return FALSE;
-
-    case WM_SYSCOMMAND:             /* 0x112 */
-    case WM_CHANGEUISTATE:          /* 0x127 */
-    case WM_UPDATEUISTATE:          /* 0x128 */
-    case WM_QUERYUISTATE:           /* 0x129*/
-    case WM_CTLCOLORBTN:            /* 0x135 */
-    case WM_CTLCOLORDLG:            /* 0x136 */
-    case WM_CTLCOLORSTATIC:         /* 0x138 */
-    case WM_MOUSEFIRST:             /* 0x200 */
-    case WM_LBUTTONDOWN:            /* 0x201 */
-    case WM_LBUTTONUP:              /* 0x202 */
-    case WM_CAPTURECHANGED:         /* 0x215 */
-    case WM_MOVING:                 /* 0x216 */
-    case WM_ENTERSIZEMOVE:          /* 0x231*/
-    case WM_EXITSIZEMOVE:           /* 0x232 */
-    case WM_IME_SETCONTEXT:         /* 0x281 */
-    case WM_IME_NOTIFY:             /* 0x282 */
-    case WM_NCMOUSELEAVE:           /* 0x2a2 */
-    case WM_PRINTCLIENT:            /* 0x318 */
-    case WM_DWMNCRENDERINGCHANGED:  /* 0x31f */
-    case WM_USER:                   /* 0x400 */
-        return FALSE;
-
-    default:  /* unexpected message type */
-        printf_s ("HelpDialog msg = %x Info1 = %lld info2 =%lld \n", message, additionalInfo1,
-            additionalInfo2);
-        return FALSE;
-    }
-
-    return FALSE;
-}
-
-/* display a dialog box so that the user can select a help topic */
-static long long helpdiag(void) {
-    hResp.radiobutton = 0;  /* set default value */
-    auto rv = DialogBoxParamW(GetModuleHandle(nullptr), MAKEINTRESOURCE(help_dialog), 
-        handConsole, helpDialogAct, (LPARAM)  99L);
-    if (rv != IDOK && rv != IDCANCEL) {
-        std::cout << "rv = " << rv << '\n';
-        ErrorDisp(__FUNCTION__);
-        return rv;
-    }
-    
-    return rv;
-}
-
-
-/*search the docfile for the required help topic. The topic can be specified in 
-the command parameter. If this is empty the topic can be selected via a dialog box;
-the required entry is specified by hResp.radiobutton. 
-Just search the file for the heading, and print everything until the next 
-heading is found.
-If the help file is not found it is possible to change the path to access the file.
-This change is retained in the BigIntCalculator.ini file. 
-The help file is not limited to ASCII characters. Most UTF-8 characters can be printed. */
-static void helpfunc(const std::vector<std::string> &command)
-{
-    FILE* doc;
-    char str[1024];
-    bool printtopic = false;
-    bool line1 = true;
-    std::string expr = " ";
-    char* newpathC;
-    std::string newpath;
-    const unsigned char BOM[] = { 0xEF, 0xBB, 0xBF };   /* byte order marker for UTF-8 */
-    bool UTF8 = false;          /* set true if UTF8 BOM found */
-    std::string helptopic;
-    int lineCount = 0;
-
-    const char butttext[][20] = { "HELP", "FUNCTION", "EXPRESSION", "OTHER",
-            "YAFU", "TEST", "MSIEVE", "BACKGROUND", "LOOP", "QMES",
-            "PARI", "AYUDA" };
-
-    if (command.size() >1) {
-        helptopic = command[1];
-    }
-    else
-        helptopic.clear();
-
-    if (lang == 0){     /* if English select one from multiple topics  */
-        if (helptopic.empty()) {  
-            helpdiag();  /* result saved in global hResp */
-            if (verbose > 0 && hResp.Ok_Cancel == IDOK) {
-                std::cout << "button = " << butttext[hResp.radiobutton] << '\n';
-            }
-            if (hResp.Ok_Cancel == IDCANCEL) {
-                std::cout << "help cancelled \n";
-                return;
-            }
-            helptopic = butttext[hResp.radiobutton];
-        }
-    }
-    else   /* español. Spanish language help is not divided into topics. */
-        helptopic = butttext[11];  /* AYUDA */
-retry:
-    //open the doc file and search for a matching topic
-    errno_t ecode = fopen_s(&doc, helpFilePath.data(), "r");
-    if (ecode != 0) {
-        /* failed to open the help file*/
-        char buffer[80];
-        _strerror_s(buffer, sizeof(buffer), NULL); /* convert errno to a text messsage */
-        fprintf_s(stderr, "fopen error: %s\n", buffer);
-        fprintf_s(stderr, "help file not found\n");
-
-        while (std::toupper(expr[0]) != 'Y') {
-            std::cout << "Do you want to search for the help file? (Y/N) \n";
-            std::getline(std::cin, expr);
-            if (std::toupper(expr[0]) == 'N')
-                return;
-        }
-        newpathC = getFileName("Text\0*.TXT\0\0", handConsole);
-        if (newpathC ==NULL) {
-            std::cout << "command cancelled \n";
-            return;
-        }
-        else {
-            helpFilePath = newpathC; /* copy new path for doc file to permanent storage */
-            writeIni();       /* update the .ini file*/
-            goto retry;       /* let's try to open the file again */
-        }
-    }
-
-    /* doc file has been opened successfully */
-    /* change stdout mode to support unicode (need to use wprintf, not printf) 
-    this allows multi-byte charactes to be printed */
-    fflush(stdout);
-    int oldmode = _setmode(_fileno(stdout), _O_U8TEXT);
-    if (verbose > 0)
-        wprintf_s(L"searching for help on '%S'\n", helptopic.c_str());
-
-    /* exit this loop when reached EOF or the next topic after the one required 
-    is reached */
-    while (!feof(doc)) {
-        
-        char *rv = fgets(str, sizeof(str), doc);   //read a line
-        if (rv == NULL)
-            if (feof(doc)) {
-                break;
-            }
-            else {
-                char buffer[80];
-                _strerror_s(buffer, sizeof(buffer), NULL); /* convert errno to a text messsage */
-                fprintf_s(stderr, "fgets error: %s\n", buffer);
-                break;
-            }
-
-        if (line1) {
-            line1 = false;     /* only do this on 1st line */
-            int d = (unsigned char)str[0] - BOM[0];    /* BOM has to be declared as unsigned char */
-            int d1 = (unsigned char)str[1] - BOM[1];
-            int d2 = (unsigned char)str[2] - BOM[2];
-            if (d == 0 && d1 == 0 && d2 == 0) {
-                memmove(str, str + 3, strlen(str) - 2);  /* remove BOM */
-                UTF8 = true;
-            }
-        }
-
-        //is this a header?
-        if ((str[0] == '[') && (str[strlen(str) - 2] == ']')) 	{
-
-            if (printtopic)
-                break;  /* we have reached the start of the next topic, so exit 
-                           Only print 1 topic per help command */
-
-            //does it match our topic?
-            str[strlen(str) - 2] = '\0'; /* overwrite ']' with null */
-            if (strstr(helptopic.c_str(), str + 1) != NULL)
-                /* we get a match if the topic between [ and ] is contained 
-                anywhere in helptopic */
-                printtopic = true;   /* we have found the required topic*/
-            lineCount = 0;    /* reset line count */
-        }
-        else {  /* not a header line */
-            if (printtopic) {
-                wprintf_s(L"%S", str);  /* print only if within the required topic */
-                lineCount++;
-            }
-            if (lineCount > 26) {
-                std::wcout << L"** More (y/n) ? ";
-                std::getline(std::cin, expr);
-                expr[0] = toupper(expr[0]);
-                if (expr[0] != 'Y')
-                    break;
-                else
-                    lineCount = 0;
-            }
-        }
-    }
-
-    if (feof(doc)) {
-        if (printtopic)
-            wprintf(L"\n");   /* contrary to the POSIX standard, the last line of the file
-                            may not end with newline */
-        else
-            if (lang)
-                wprintf_s(L"Ayuda para %S no encontrado \n", helptopic.c_str());
-            else
-                wprintf_s(L"Help for %S not found \n", helptopic.c_str());
-    }
-    /* change stdout back to normal */
-    fflush(stdout);
-    _setmode(_fileno(stdout), oldmode);
-    fclose(doc);
-    return;
-}
-
-
-/* code 'borrowed' from YAFU */
-// machine info
-double MEAS_CPU_FREQUENCY;
-char CPU_ID_STR[80] = {'\0'};
-
-int CLSIZE;
-char HAS_SSE41;
-#if defined(WIN32)
-char sysname[MAX_COMPUTERNAME_LENGTH + 1];
-unsigned long sysname_sz;
-#else
-char sysname[256];
-int sysname_sz;
-#endif
-
-#ifdef _MSC_VER
-
-/* Core aware timing on Windows, courtesy of Brian Gladman */
-
-#if defined( _WIN64 )
-
-#define current_processor_number GetCurrentProcessorNumber
-
-#else
-
-unsigned long current_processor_number(void)
-{
-    __asm
-    {
-        mov     eax, 1
-        cpuid
-        shr     ebx, 24
-        mov     eax, ebx
-    }
-}
-
-#endif
-
-static int lock_thread_to_core(void)
-{
-    DWORD_PTR afp, afs;
-
-    if (GetProcessAffinityMask(GetCurrentProcess(), &afp, &afs))
-    {
-        afp &= (DWORD_PTR)(1LL << current_processor_number());
-        if (SetThreadAffinityMask(GetCurrentThread(), afp))
-            return EXIT_SUCCESS;
-    }
-    return EXIT_FAILURE;
-}
-
-static int unlock_thread_from_core(void)
-{
-    DWORD_PTR afp, afs;
-
-    if (GetProcessAffinityMask(GetCurrentProcess(), &afp, &afs))
-    {
-        if (SetThreadAffinityMask(GetCurrentThread(), afp))
-            return EXIT_SUCCESS;
-    }
-    return EXIT_FAILURE;
-}
-
-double cycles_per_second = 0.0;
-double ticks_per_second = 0.0;
-double cycles_per_tick = 0.0;
-
-static unsigned long long measure_processor_speed(void)
-{
-    unsigned long long cycles;
-
-    lock_thread_to_core();
-    cycles = __rdtsc();
-    Sleep(100);
-    cycles = __rdtsc() - cycles;
-    unlock_thread_from_core();
-    cycles_per_second = 10.0 * (double)cycles;
-
-    if (ticks_per_second == 0.0)
-    {
-        LARGE_INTEGER ll;
-        QueryPerformanceFrequency(&ll);
-        ticks_per_second = (double)ll.QuadPart;
-        cycles_per_tick = cycles_per_second / ticks_per_second;
-    }
-    return cycles;
-}
-
-
-#else
-
-uint64 measure_processor_speed(void)
-{
-    uint64 cycles;
-    struct timeval start, stop;
-    double t_time;
-    TIME_DIFF* difference;
-
-    gettimeofday(&start, NULL);
-
-    cycles = yafu_read_clock();
-    do
-    {
-        gettimeofday(&stop, NULL);
-        difference = my_difftime(&start, &stop);
-        t_time = ((double)difference->secs +
-            (double)difference->usecs / 1000000);
-        free(difference);
-    } while (t_time < 0.1);
-    cycles = yafu_read_clock() - cycles;
-
-    return cycles;                  /* return cycles per second  */
-}
-
-#endif
-
-#if defined(GCC_ASM32X)
-#define HAS_CPUID
-#define CPUID(code, a, b, c, d) 			\
-        ASM_G volatile(					\
-            "movl %%ebx, %%esi   \n\t"		\
-            "cpuid               \n\t"		\
-            "movl %%ebx, %1      \n\t"		\
-            "movl %%esi, %%ebx   \n\t"		\
-            :"=a"(a), "=m"(b), "=c"(c), "=d"(d) 	\
-            :"0"(code) : "%esi")
-#define CPUID2(code1, code2, a, b, c, d) 			\
-        ASM_G volatile(					\
-            "movl %%ebx, %%esi   \n\t"		\
-            "cpuid               \n\t"		\
-            "movl %%ebx, %1      \n\t"		\
-            "movl %%esi, %%ebx   \n\t"		\
-            :"=a"(a), "=m"(b), "=c"(c), "=d"(d) 	\
-            :"0"(code1), "2"(code2) : "%esi")
-
-#elif defined(GCC_ASM64X)
-#define HAS_CPUID
-#define CPUID(code, a, b, c, d) 			\
-        ASM_G volatile(					\
-            "movq %%rbx, %%rsi   \n\t"		\
-            "cpuid               \n\t"		\
-            "movl %%ebx, %1      \n\t"		\
-            "movq %%rsi, %%rbx   \n\t"		\
-            :"=a"(a), "=m"(b), "=c"(c), "=d"(d) 	\
-            :"0"(code) : "%rsi")
-#define CPUID2(code1, code2, a, b, c, d)		\
-        ASM_G volatile(					\
-            "movq %%rbx, %%rsi   \n\t"		\
-            "cpuid               \n\t"		\
-            "movl %%ebx, %1      \n\t"		\
-            "movq %%rsi, %%rbx   \n\t"		\
-            :"=a"(a), "=m"(b), "=c"(c), "=d"(d) 	\
-            :"0"(code1), "2"(code2) : "%rsi")
-
-#elif defined(_MSC_VER)
-#include <intrin.h>
-#define HAS_CPUID
-#define CPUID(code, a, b, c, d)	\
-    {	int _z[4]; \
-        __cpuid(_z, code); \
-        a = _z[0]; \
-        b = _z[1]; \
-        c = _z[2]; \
-        d = _z[3]; \
-    }
-#define CPUID2(code1, code2, a, b, c, d) \
-    {	int _z[4]; \
-        __cpuidex(_z, code1, code2); \
-        a = _z[0]; \
-        b = _z[1]; \
-        c = _z[2]; \
-        d = _z[3]; \
-    }
-#endif
-
-
-// http://msdn.microsoft.com/en-us/library/hskdteyh.aspx
-// cpuid.cpp 
-// processor: x86, x64
-// Use the __cpuid intrinsic to get information about a CPU
-// modified for c compliers and use of CPUID macros 
-//		- brb, 10/26/10
-
-const char* szFeatures[] =
-{
-    "x87 FPU On Chip",
-    "Virtual-8086 Mode Enhancement",
-    "Debugging Extensions",
-    "Page Size Extensions",
-    "Time Stamp Counter",
-    "RDMSR and WRMSR Support",
-    "Physical Address Extensions",
-    "Machine Check Exception",
-    "CMPXCHG8B Instruction",
-    "APIC On Chip",
-    "Unknown1",
-    "SYSENTER and SYSEXIT",
-    "Memory Type Range Registers",
-    "PTE Global Bit",
-    "Machine Check Architecture",
-    "Conditional Move/Compare Instruction",
-    "Page Attribute Table",
-    "36-bit Page Size Extension",
-    "Processor Serial Number",
-    "CFLUSH Extension",
-    "Unknown2",
-    "Debug Store",
-    "Thermal Monitor and Clock Ctrl",
-    "MMX Technology",
-    "FXSAVE/FXRSTOR",
-    "SSE Extensions",
-    "SSE2 Extensions",
-    "Self Snoop",
-    "Multithreading Technology",
-    "Thermal Monitor",
-    "Unknown4",
-    "Pending Break Enable"
-};
-
-
-
-/* return cpuid , cache size, etc  */
-static int extended_cpuid(char* CPUidstr, int* cachelinesize, char* bSSE41Extensions, 
-    int do_print)
-{
-    char CPUString[0x20];
-    char CPUBrandString[0x40] = { 0 };
-    int CPUInfo[4] = { -1 };
-    int nSteppingID = 0;
-    int nModel = 0;
-    int nFamily = 0;
-    int nProcessorType = 0;
-    int nExtendedmodel = 0;
-    int nExtendedfamily = 0;
-    int nBrandIndex = 0;
-    int nCLFLUSHcachelinesize = 0;
-    int nLogicalProcessors = 0;
-    int nAPICPhysicalID = 0;
-    int nFeatureInfo = 0;
-    int nCacheLineSize = 0;
-    int nL2Associativity = 0;
-    int nCacheSizeK = 0;
-    int nPhysicalAddress = 0;
-    int nVirtualAddress = 0;
-    int nRet = 0;    /* return value */
-
-    int nCores = 0;
-    int nCacheType = 0;
-    int nCacheLevel = 0;
-    int nMaxThread = 0;
-    int nSysLineSize = 0;
-    int nPhysicalLinePartitions = 0;
-    int nWaysAssociativity = 0;
-    int nNumberSets = 0;
-
-    unsigned    nIds, nExIds, i;
-
-    char    bSSE3Instructions = 0;
-    char    bMONITOR_MWAIT = 0;
-    char    bCPLQualifiedDebugStore = 0;
-    char    bVirtualMachineExtensions = 0;
-    char    bEnhancedIntelSpeedStepTechnology = 0;
-    char    bThermalMonitor2 = 0;
-    char    bSupplementalSSE3 = 0;
-    char    bL1ContextID = 0;
-    char    bCMPXCHG16B = 0;
-    char    bxTPRUpdateControl = 0;
-    char    bPerfDebugCapabilityMSR = 0;
-    //char    bSSE41Extensions = 0;
-    char    bSSE42Extensions = 0;
-    char    bPOPCNT = 0;
-
-    char    bMultithreading = 0;
-
-    char    bLAHF_SAHFAvailable = 0;
-    char    bCmpLegacy = 0;
-    char    bSVM = 0;
-    char    bExtApicSpace = 0;
-    char    bAltMovCr8 = 0;
-    char    bLZCNT = 0;
-    char    bSSE4A = 0;
-    char    bMisalignedSSE = 0;
-    char    bPREFETCH = 0;
-    char    bSKINITandDEV = 0;
-    char    bSYSCALL_SYSRETAvailable = 0;
-    char    bExecuteDisableBitAvailable = 0;
-    char    bMMXExtensions = 0;
-    char    bFFXSR = 0;
-    char    b1GBSupport = 0;
-    char    bRDTSCP = 0;
-    char    b64Available = 0;
-    char    b3DNowExt = 0;
-    char    b3DNow = 0;
-    char    bNestedPaging = 0;
-    char    bLBRVisualization = 0;
-    char    bFP128 = 0;
-    char    bMOVOptimization = 0;
-
-    char    bSelfInit = 0;
-    char    bFullyAssociative = 0;
-
-    *bSSE41Extensions = 0;
-
-    // __cpuid with an InfoType argument of 0 returns the number of
-    // valid Ids in CPUInfo[0] and the CPU identification string in
-    // the other three array elements. The CPU identification string is
-    // not in linear order. The code below arranges the information 
-    // in a human readable form.
-    CPUID(0, CPUInfo[0], CPUInfo[1], CPUInfo[2], CPUInfo[3]);
-    //__cpuid(CPUInfo, 0);
-    nIds = CPUInfo[0];
-    memset(CPUString, 0, sizeof(CPUString));
-    *((int*)CPUString) = CPUInfo[1];
-    *((int*)(CPUString + 4)) = CPUInfo[3];
-    *((int*)(CPUString + 8)) = CPUInfo[2];
-
-    // Get the information associated with each valid Id
-    for (i = 0; i <= nIds; ++i)
-    {
-        CPUID(i, CPUInfo[0], CPUInfo[1], CPUInfo[2], CPUInfo[3]);
-        //__cpuid(CPUInfo, i);
-
-        if (do_print > 1)
-        {
-            printf("\nFor InfoType %d\n", i);
-            printf("CPUInfo[0] = 0x%x\n", CPUInfo[0]);
-            printf("CPUInfo[1] = 0x%x\n", CPUInfo[1]);
-            printf("CPUInfo[2] = 0x%x\n", CPUInfo[2]);
-            printf("CPUInfo[3] = 0x%x\n", CPUInfo[3]);
-        }
-
-        // Interpret CPU feature information.
-        if (i == 1)
-        {
-            nSteppingID = CPUInfo[0] & 0xf;
-            nModel = (CPUInfo[0] >> 4) & 0xf;
-            nFamily = (CPUInfo[0] >> 8) & 0xf;
-            nProcessorType = (CPUInfo[0] >> 12) & 0x3;
-            nExtendedmodel = (CPUInfo[0] >> 16) & 0xf;
-            nExtendedfamily = (CPUInfo[0] >> 20) & 0xff;
-            nBrandIndex = CPUInfo[1] & 0xff;
-            *cachelinesize = nCLFLUSHcachelinesize = ((CPUInfo[1] >> 8) & 0xff) * 8;
-            nLogicalProcessors = ((CPUInfo[1] >> 16) & 0xff);
-            nAPICPhysicalID = (CPUInfo[1] >> 24) & 0xff;
-            bSSE3Instructions = (CPUInfo[2] & 0x1) || 0;
-            bMONITOR_MWAIT = (CPUInfo[2] & 0x8) || 0;
-            bCPLQualifiedDebugStore = (CPUInfo[2] & 0x10) || 0;
-            bVirtualMachineExtensions = (CPUInfo[2] & 0x20) || 0;
-            bEnhancedIntelSpeedStepTechnology = (CPUInfo[2] & 0x80) || 0;
-            bThermalMonitor2 = (CPUInfo[2] & 0x100) || 0;
-            bSupplementalSSE3 = (CPUInfo[2] & 0x200) || 0;
-            bL1ContextID = (CPUInfo[2] & 0x300) || 0;
-            bCMPXCHG16B = (CPUInfo[2] & 0x2000) || 0;
-            bxTPRUpdateControl = (CPUInfo[2] & 0x4000) || 0;
-            bPerfDebugCapabilityMSR = (CPUInfo[2] & 0x8000) || 0;
-            *bSSE41Extensions = (CPUInfo[2] & 0x80000) || 0;
-            bSSE42Extensions = (CPUInfo[2] & 0x100000) || 0;
-            bPOPCNT = (CPUInfo[2] & 0x800000) || 0;
-            nFeatureInfo = CPUInfo[3];
-            bMultithreading = (nFeatureInfo & (1 << 28)) || 0;
-        }
-    }
-
-    // Calling __cpuid with 0x80000000 as the InfoType argument
-    // gets the number of valid extended IDs.
-    CPUID(0x80000000, CPUInfo[0], CPUInfo[1], CPUInfo[2], CPUInfo[3]);
-    //__cpuid(CPUInfo, 0x80000000);
-    nExIds = CPUInfo[0];
-    memset(CPUBrandString, 0, sizeof(CPUBrandString));
-
-    // Get the information associated with each extended ID.
-    for (i = 0x80000000; i <= nExIds; ++i)
-    {
-        CPUID(i, CPUInfo[0], CPUInfo[1], CPUInfo[2], CPUInfo[3]);
-        //__cpuid(CPUInfo, i);
-        if (do_print > 1)
-        {
-            printf("\nFor InfoType %x\n", i);
-            printf("CPUInfo[0] = 0x%x\n", CPUInfo[0]);
-            printf("CPUInfo[1] = 0x%x\n", CPUInfo[1]);
-            printf("CPUInfo[2] = 0x%x\n", CPUInfo[2]);
-            printf("CPUInfo[3] = 0x%x\n", CPUInfo[3]);
-        }
-
-        if (i == 0x80000001)
-        {
-            bLAHF_SAHFAvailable = (CPUInfo[2] & 0x1) || 0;
-            bCmpLegacy = (CPUInfo[2] & 0x2) || 0;
-            bSVM = (CPUInfo[2] & 0x4) || 0;
-            bExtApicSpace = (CPUInfo[2] & 0x8) || 0;
-            bAltMovCr8 = (CPUInfo[2] & 0x10) || 0;
-            bLZCNT = (CPUInfo[2] & 0x20) || 0;
-            bSSE4A = (CPUInfo[2] & 0x40) || 0;
-            bMisalignedSSE = (CPUInfo[2] & 0x80) || 0;
-            bPREFETCH = (CPUInfo[2] & 0x100) || 0;
-            bSKINITandDEV = (CPUInfo[2] & 0x1000) || 0;
-            bSYSCALL_SYSRETAvailable = (CPUInfo[3] & 0x800) || 0;
-            bExecuteDisableBitAvailable = (CPUInfo[3] & 0x10000) || 0;
-            bMMXExtensions = (CPUInfo[3] & 0x40000) || 0;
-            bFFXSR = (CPUInfo[3] & 0x200000) || 0;
-            b1GBSupport = (CPUInfo[3] & 0x400000) || 0;
-            bRDTSCP = (CPUInfo[3] & 0x8000000) || 0;
-            b64Available = (CPUInfo[3] & 0x20000000) || 0;
-            b3DNowExt = (CPUInfo[3] & 0x40000000) || 0;
-            b3DNow = (CPUInfo[3] & 0x80000000) || 0;
-        }
-
-        // Interpret CPU brand string and cache information.
-        if (i == 0x80000002)
-            memcpy(CPUBrandString, CPUInfo, sizeof(CPUInfo));
-        else if (i == 0x80000003)
-            memcpy(CPUBrandString + 16, CPUInfo, sizeof(CPUInfo));
-        else if (i == 0x80000004)
-            memcpy(CPUBrandString + 32, CPUInfo, sizeof(CPUInfo));
-        else if (i == 0x80000006)
-        {
-            nCacheLineSize = CPUInfo[2] & 0xff;
-            nL2Associativity = (CPUInfo[2] >> 12) & 0xf;
-            nCacheSizeK = (CPUInfo[2] >> 16) & 0xffff;
-        }
-        else if (i == 0x80000008)
-        {
-            nPhysicalAddress = CPUInfo[0] & 0xff;
-            nVirtualAddress = (CPUInfo[0] >> 8) & 0xff;
-        }
-        else if (i == 0x8000000A)
-        {
-            bNestedPaging = (CPUInfo[3] & 0x1) || 0;
-            bLBRVisualization = (CPUInfo[3] & 0x2) || 0;
-        }
-        else if (i == 0x8000001A)
-        {
-            bFP128 = (CPUInfo[0] & 0x1) || 0;
-            bMOVOptimization = (CPUInfo[0] & 0x2) || 0;
-        }
-    }
-    ptrdiff_t ix = 0;
-    while (ix < 64 && isblank(CPUBrandString[ix]))
-        ix++;        /* find 1st non-blank character */
-    strcpy_s(CPUidstr, 64 - ix, CPUBrandString + ix);
-    // Display all the information in user-friendly format.
-    if (do_print > 0)
-        printf("\n\nCPU String: %s\n", CPUString);
-
-    if (nIds >= 1)
-    {
-        if (do_print > 0)
-        {
-            if (nSteppingID)
-                printf("Stepping ID = %d\n", nSteppingID);
-            if (nModel)
-                printf("Model = %d\n", nModel);
-            if (nFamily)
-                printf("Family = %d\n", nFamily);
-            if (nProcessorType)
-                printf("Processor Type = %d\n", nProcessorType);
-            if (nExtendedmodel)
-                printf("Extended model = %d\n", nExtendedmodel);
-            if (nExtendedfamily)
-                printf("Extended family = %d\n", nExtendedfamily);
-            if (nBrandIndex)
-                printf("Brand Index = %d\n", nBrandIndex);
-            if (nCLFLUSHcachelinesize)
-                printf("CLFLUSH cache line size = %d\n",
-                    nCLFLUSHcachelinesize);
-            if (bMultithreading && (nLogicalProcessors > 0))
-                printf("Logical Processor Count = %d\n", nLogicalProcessors);
-            if (nAPICPhysicalID)
-                printf("APIC Physical ID = %d\n", nAPICPhysicalID);
-
-            if (nFeatureInfo || bSSE3Instructions ||
-                bMONITOR_MWAIT || bCPLQualifiedDebugStore ||
-                bVirtualMachineExtensions || bEnhancedIntelSpeedStepTechnology ||
-                bThermalMonitor2 || bSupplementalSSE3 || bL1ContextID ||
-                bCMPXCHG16B || bxTPRUpdateControl || bPerfDebugCapabilityMSR ||
-                *bSSE41Extensions || bSSE42Extensions || bPOPCNT ||
-                bLAHF_SAHFAvailable || bCmpLegacy || bSVM ||
-                bExtApicSpace || bAltMovCr8 ||
-                bLZCNT || bSSE4A || bMisalignedSSE ||
-                bPREFETCH || bSKINITandDEV || bSYSCALL_SYSRETAvailable ||
-                bExecuteDisableBitAvailable || bMMXExtensions || bFFXSR || b1GBSupport ||
-                bRDTSCP || b64Available || b3DNowExt || b3DNow || bNestedPaging ||
-                bLBRVisualization || bFP128 || bMOVOptimization)
-            {
-                printf("\nThe following features are supported:\n");
-
-                if (bSSE3Instructions)
-                    printf("\tSSE3\n");
-                if (bMONITOR_MWAIT)
-                    printf("\tMONITOR/MWAIT\n");
-                if (bCPLQualifiedDebugStore)
-                    printf("\tCPL Qualified Debug Store\n");
-                if (bVirtualMachineExtensions)
-                    printf("\tVirtual Machine Extensions\n");
-                if (bEnhancedIntelSpeedStepTechnology)
-                    printf("\tEnhanced Intel SpeedStep Technology\n");
-                if (bThermalMonitor2)
-                    printf("\tThermal Monitor 2\n");
-                if (bSupplementalSSE3)
-                    printf("\tSupplemental Streaming SIMD Extensions 3\n");
-                if (bL1ContextID)
-                    printf("\tL1 Context ID\n");
-                if (bCMPXCHG16B)
-                    printf("\tCMPXCHG16B Instruction\n");
-                if (bxTPRUpdateControl)
-                    printf("\txTPR Update Control\n");
-                if (bPerfDebugCapabilityMSR)
-                    printf("\tPerf\\Debug Capability MSR\n");
-                if (*bSSE41Extensions)
-                    printf("\tSSE4.1 Extensions\n");
-                if (bSSE42Extensions)
-                    printf("\tSSE4.2 Extensions\n");
-                if (bPOPCNT)
-                    printf("\tPPOPCNT Instruction\n");
-
-                i = 0;
-                nIds = 1;
-                while (i < (sizeof(szFeatures) / sizeof(const char*)))
-                {
-                    if (nFeatureInfo & nIds)
-                    {
-                        printf("\t");
-                        printf("%s", szFeatures[i]);
-                        printf("\n");
-                    }
-
-                    nIds <<= 1;
-                    ++i;
-                }
-                if (bLAHF_SAHFAvailable)
-                    printf("\tLAHF/SAHF in 64-bit mode\n");
-                if (bCmpLegacy)
-                    printf("\tCore multi-processing legacy mode\n");
-                if (bSVM)
-                    printf("\tSecure Virtual Machine\n");
-                if (bExtApicSpace)
-                    printf("\tExtended APIC Register Space\n");
-                if (bAltMovCr8)
-                    printf("\tAltMovCr8\n");
-                if (bLZCNT)
-                    printf("\tLZCNT instruction\n");
-                if (bSSE4A)
-                    printf("\tSSE4A (EXTRQ, INSERTQ, MOVNTSD, MOVNTSS)\n");
-                if (bMisalignedSSE)
-                    printf("\tMisaligned SSE mode\n");
-                if (bPREFETCH)
-                    printf("\tPREFETCH and PREFETCHW Instructions\n");
-                if (bSKINITandDEV)
-                    printf("\tSKINIT and DEV support\n");
-                if (bSYSCALL_SYSRETAvailable)
-                    printf("\tSYSCALL/SYSRET in 64-bit mode\n");
-                if (bExecuteDisableBitAvailable)
-                    printf("\tExecute Disable Bit\n");
-                if (bMMXExtensions)
-                    printf("\tExtensions to MMX Instructions\n");
-                if (bFFXSR)
-                    printf("\tFFXSR\n");
-                if (b1GBSupport)
-                    printf("\t1GB page support\n");
-                if (bRDTSCP)
-                    printf("\tRDTSCP instruction\n");
-                if (b64Available)
-                    printf("\t64 bit Technology\n");
-                if (b3DNowExt)
-                    printf("\t3Dnow Ext\n");
-                if (b3DNow)
-                    printf("\t3Dnow! instructions\n");
-                if (bNestedPaging)
-                    printf("\tNested Paging\n");
-                if (bLBRVisualization)
-                    printf("\tLBR Visualization\n");
-                if (bFP128)
-                    printf("\tFP128 optimization\n");
-                if (bMOVOptimization)
-                    printf("\tMOVU Optimization\n");
-            }
-        }
-    }
-
-    if (nExIds >= 0x80000004 && do_print > 0)
-        printf("\nCPU Brand String: %s\n", CPUBrandString);
-
-    if (nExIds >= 0x80000006 && do_print >0)
-    {
-        printf("Cache Line Size = %d\n", nCacheLineSize);
-        printf("L2 Associativity = %d\n", nL2Associativity);
-        printf("Cache Size = %dK\n", nCacheSizeK);
-    }
-
-
-    for (i = 0;; i++)
-    {
-        CPUID2(0x4, i, CPUInfo[0], CPUInfo[1], CPUInfo[2], CPUInfo[3]);
-        //__cpuidex(CPUInfo, 0x4, i);
-        if (!(CPUInfo[0] & 0xf0)) 
-            break;
-
-        if (i == 0) {
-            nCores = CPUInfo[0] >> 26;
-            if (do_print > 0)
-                printf("\n\nNumber of Cores = %d\n", nCores + 1);
-        }
-
-        nCacheType = (CPUInfo[0] & 0x1f);
-        nCacheLevel = (CPUInfo[0] & 0xe0) >> 5;
-        bSelfInit = (CPUInfo[0] & 0x100) >> 8;
-        bFullyAssociative = (CPUInfo[0] & 0x200) >> 9;
-        nMaxThread = (CPUInfo[0] & 0x03ffc000) >> 14;
-        nSysLineSize = (CPUInfo[1] & 0x0fff);
-        nPhysicalLinePartitions = (CPUInfo[1] & 0x03ff000) >> 12;
-        nWaysAssociativity = (CPUInfo[1]) >> 22;
-        nNumberSets = CPUInfo[2];
-
-        if (do_print > 0) {
-            printf("\n");
-
-            printf("ECX Index %d\n", i);
-            switch (nCacheType)
-            {
-            case 0:
-                printf("   Type: Null\n");
-                break;
-            case 1:
-                printf("   Type: Data Cache\n");
-                break;
-            case 2:
-                printf("   Type: Instruction Cache\n");
-                break;
-            case 3:
-                printf("   Type: Unified Cache\n");
-                break;
-            default:
-                printf("   Type: Unknown\n");
-            }
-
-            printf("   Level = %d\n", nCacheLevel + 1);
-            if (bSelfInit) 	{
-                printf("   Self Initializing\n");
-            }
-            else {
-                printf("   Not Self Initializing\n");
-            }
-            if (bFullyAssociative) {
-                printf("   Is Fully Associatve\n");
-            }
-            else {
-                printf("   Is Not Fully Associatve\n");
-            }
-            printf("   Max Threads = %d\n", nMaxThread + 1);
-            printf("   System Line Size = %d\n", nSysLineSize + 1);
-            printf("   Physical Line Partions = %d\n", nPhysicalLinePartitions + 1);
-            printf("   Ways of Associativity = %d\n", nWaysAssociativity + 1);
-            printf("   Number of Sets = %d\n", nNumberSets + 1);
-        }
-    }
-
-    return  nRet;
-}
-
-// function containing system commands to get the computer name, CPU speed, etc
-static void get_computer_info(char* CPUidstr)
-{
-
-    //figure out cpu freq. 0.1 seconds won't be very accurate 
-    MEAS_CPU_FREQUENCY = measure_processor_speed() / 1.0e5;
-
-#ifdef __APPLE__
-    // something in extended cpuid causes a segfault on mac builds.
-    // just disable it for now - this information is not critical for
-    // program operation.
-    strcpy(idstr, "N/A");
-    CLSIZE = 0;
-    L1CACHE = DEFAULT_L1_CACHE_SIZE;
-    L2CACHE = DEFAULT_L2_CACHE_SIZE;
-    HAS_SSE41 = 0;
-
-#else
-    
-    // run an extended cpuid command to get the cache line size, and
-    // optionally print a bunch of info to the screen
-
-    extended_cpuid(CPUidstr, &CLSIZE, &HAS_SSE41, verbose);
-#endif
-
-#if defined(WIN32)
-
-    sysname_sz = MAX_COMPUTERNAME_LENGTH + 1;
-    GetComputerNameA(sysname, &sysname_sz);
-
-#else
-
-    ret = gethostname(sysname, sizeof(sysname) / sizeof(*sysname));
-    sysname[(sizeof(sysname) - 1) / sizeof(*sysname)] = 0;	// null terminate
-    if (ret != 0)
-    {
-        printf("error occured when getting host name\n");
-        strcpy(sysname, "N/A");
-    }
-    sysname_sz = strlen(sysname);
-
-#endif
-
-    return;
-}
-
-
-/* evaluate 1 or more expressions, separated by commas */
-static retCode ComputeMultiExpr(std::string expr, Znum result) {
-    std::string subExpr;
-    retCode rv = retCode::EXPR_OK;
-    size_t subStart = 0, subEnd;
-    int bc = 0;   /* bracket count */
-    int asgCt = 0;   /* number of assignment operators */
-    while (subStart < expr.size()) {
-        bc = 0;
-        /* find  separator or end of text*/
-        for (subEnd = subStart + 1; subEnd < expr.size(); subEnd++) {
-            if (expr[subEnd] == '(')
-                bc++;
-            if (expr[subEnd] == ')')
-                bc--;
-            if (bc == 0 && expr[subEnd] == ',')
-                break;
-        }
-        if (bc != 0)
-            return retCode::PAREN_MISMATCH;
-        subExpr = expr.substr(subStart, subEnd - subStart);
-        removeInitTrail(subExpr);  /* remove initial & trailing blanks */
-        removeIntSpace(subExpr);   /* remove spaces between digits */
-        rv = ComputeExpr(subExpr, result, asgCt);
-        if (rv != retCode::EXPR_OK) {
-            textError(rv);   // invalid expression; print error message
-            return rv;
-        }
-        else {
-            if (asgCt == 0)
-                std::cout << " = ";
-            else {
-                /* print names of variables assigned values */
-                for (size_t ix = 0; ix < subExpr.size() && asgCt > 0; ix++) {
-                    putchar(subExpr[ix]);
-                    if (subExpr[ix] == '=')
-                        asgCt--;
-                }
-            }
-
-            ShowLargeNumber(result, groupSize, true, hexPrFlag);   // print value of expression
-            std::cout << '\n';
-
-        }
-        subStart = subEnd + 1;  /* move past , */
-    } /* end of while loop */
-
-    return rv;
-}
-
 /* format is IF (expression) REPEAT 
           or IF (expression) STOP 
           or IF (expression) THEN (expression) [ELSE (expression)]
@@ -4180,7 +2947,7 @@ the _MSC_FULL_VER macro evaluates to 150020706 */
     processIni(argv[0]); // read .ini file if it exists
 
     //get the computer name, cache sizes, etc.  store in globals
-    get_computer_info(CPU_ID_STR);
+    get_computer_info(CPU_ID_STR, MEAS_CPU_FREQUENCY);
 
     printf_s("detected %s\n", CPU_ID_STR);
     printf_s("measured cpu frequency ~= %.0f\n", 	MEAS_CPU_FREQUENCY);
